Fixing implicit conversion warnings.

+++ src/Networking.h
@@ -164,7 +164,7 @@ private:
     static int passwordCallback(char *buf, int size, int rwflag, void *u)
     {
         std::string *password = (std::string *) u;
-        int length = std::min<int>(size, password->length());
+        int length = std::min<int>(size, (int) password->length());
         memcpy(buf, password->data(), length);
         buf[length] = '\0';
         return length;
@@ -216,7 +216,7 @@ struct WIN32_EXPORT NodeData {
     static void asyncCallback(Async *async);
 
     static int getMemoryBlockIndex(size_t length) {
-        return (length >> 4) + bool(length & 15);
+        return (int) (length >> 4) + bool(length & 15);
     }
 
     char *getSmallMemoryBlock(int index) {
